server/requirements.txt
aiohttp
aiofiles
pydantic
fastapi
uvicorn
biopython
huggingface
numpy==1.26.4
transformers==4.41.0
mdtraj